feat(booster): replace old booster system with new provider-based system
Removed:
- Old booster system (Booster interface, BoosterEvent, Boosters class,
  and related events and implementations)

Added:
- New BoosterManager, BoosterProvider, and Booster interface in the
  `api.booster` package
- SimpleBooster implementation

Changed:
- IVoidChest.booster() now returns BigDecimal (was Double)
- Removed booster methods from IPlayerData

The new system uses a provider-based approach for managing boosters.

+++ src/main/java/com/georgev22/voidchest/api/storage/data/IVoidChest.java
@@ -13,6 +13,7 @@ import com.georgev22.voidchest.api.utilities.SerializableLocation;
 import org.bukkit.inventory.Inventory;
 import org.jetbrains.annotations.*;
 
+import java.math.BigDecimal;
 import java.util.List;
 import java.util.UUID;
 
@@ -153,7 +154,7 @@ public interface IVoidChest extends Entity {
      *
      * @return The booster value of the VoidChest.
      */
-    Double booster();
+    BigDecimal booster();
 
     /**
      * Sets the booster value of the VoidChest.
